pool creation fee handled in gamm module
